backend/requirements.txt
Django==5.0.6
djangorestframework==3.16.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
Pillow
dj-database-url